lab4/image-blur-app/requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
google-cloud-storage==1.42.3
Pillow==9.5.0